engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-analyst"
version = "0.1.0"
description = "AI-powered financial analysis automation engine"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.12.0",
    "feedparser>=6.0.10",
    "pydantic>=2.6.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "yfinance>=0.2.36",
    "fredapi>=0.5.2",
    "pandas>=2.2.0",
    "xlsxwriter>=3.2.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.1.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "APScheduler>=3.10,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

[project.scripts]
analyst = "analyst.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
